Changed model settings so model would simulate correctly
Changed the fixed step size in the model settings for SimplePulseGenerator_4x4_HDL_IQ.slx to auto from 1/fpga_clk_rate. The project will run in the hardware with either but will not simulate correctly with the latter. I changed this at one point earlier to attempt to correct the time scales on the various plots in the simulation. Tony mentioned that we could modify the settings to correct these. The change to 1/fpga_clk_rate  did not appear to fix that an made the simulation run slowly or not at all. Also changed the PRF to 10000Hz which is more realistic for simulation although the hardware can only support around 10Hz while sending IQ data across the 1G Ethernet link real time. Change back to 10Hz to run in hardware.