Corrected calculation of start_inc and end_inc
These calculations were not correct in the previous version, updated in model_init.m. I don't think there is a significant change in simulink model even though it shows up as modified. I changed the counter type attempting to solve the initial count from 0 for first pulse but couldn't get that to work so changed back to free running.