Added Time scope to investigate max frequency NCO's can support
For a basic NCO, the clock rate limits the maximum frequency that can be synthesized if I understand the operation correctly. I think that same limit applies to the 4 NCO's in this design and we are limited to chirp bandwidths less than 64MHz with this approach. I added a time scope with all 4 NCO outputs, hdl_counter, and PRF_rst. I set this up to ramp from 1MHz to 128MHz, the ramp is fairly well behaved to 64MHz. Beyond that this approach does not appear to work.