LFS setup for system_wrapper.bit in .gitignore and .gitattributes
Setup .gitignore and .gitattributes to track
hdl_prj/vivado_ip_prj/vivado_prj.runs/impl_1/system_wrapper.bit
using the LFS (Large File Storage).

This file isn't yet tracked, but changes will be displayed, and if it
is added and committed, it will be stored into LFS instead of
the native Git Repo.

On the CLI, "git lfs install" is all that is required to setup the
repo to work with LFS.
https://git-lfs.github.com/