For NCO simulates, but with incorrect output. Assn init modified
Modified the Assignment block inside the For Iterator Subsystem
to NOT take an initialization input, and instead allow the
dimensions of each output port to be specified in a table.
This seems to indicate that my initialization input should have
been a muxed vector already, because the output dims are
derived from it.
How do I create a complex(0,0) with dim (4) ?

Still, the data output in "Spectrum 4xNCO Muxed in For Iterator"
isn't correct, so it isn't being treated as a MUXed input, although it
does see that the Sample rate is 512 MHz.