Add ability to load LFM counter step size
Modified masked HDL Counter to allow loading of the step size. Did not make a masked subsystem, implemented as a simple subsystem with the counter parameters in model_init so I could focus on getting the frequency response correct. The simulink simulation seems perfect, running on the hardware the spectrum returned with the HostIO interface seems flipped but otherwise usually looks reasonable. One exception is that using a value of zero for start or end frequency does not work in hardware. Also changed the data types for the LFM counter parameters, seems to pass those values correctly now. Mixers must not be working like I thought, the spectrum analyzer connected to zcu111 does not match HostIO output.