Fixed Debug  / ADC input select logic in AXI4Stream_DMA_S2MM
This logic was reversed selecting Debug input when normal operation was selected and normal input when Debug mode was selected. This resulted in no output on the time scope or spectrum analyzer in this component for final output. Debug mode just loops back in the FPGA logic so this option is available in the hardware rather than going out to the simulation environment and looping back in the simulation.