Modified Vectorized NCO For Each for Synthesis
So far, employing a for loop block in HDL code generation has
been a challenge.  The key change here is removing
partitioned parameters, and rather making the parameters
into a new explicit port on the For Each subsystem.  The
vectorized input to the For Each subsystem were generated by the
enclosing block's mask initialization.

Additionally, the DSP Flip block is not supported by HDL Coder,
so this has been replaced by a Matlab Function block.