Matched delay between pulse control logic and output mux
From the simulation there is 7 clocks of latency from the time the counter logic starts until the NCO output arrives at the mux to turn off the output between pulses. This was resulting in pulses that wrapped in frequency  and did other odd things. Fixed by adding a delay in the gating logic, this will have to be taken into account when calculating delays for the returns.